refactor: move AbstractPydanticJSONBModel to common/base_pydantic_jsonb.py
Remove misleading mixins/ folder — class is an abstract base model, not a mixin.
Update all imports accordingly.
