+++ chord_metadata_service/chord/models.py
@@ -12,7 +12,7 @@ from chord_metadata_service.resources.models import Resource
 from chord_metadata_service.restapi.schema_ref import SchemaRefs
 from chord_metadata_service.restapi.validators import JsonSchemaValidator
 from chord_metadata_service.restapi.models import BaseTimeStamp, SchemaType
-from chord_metadata_service.common.mixins.pydantic_mixin import AbstractPydanticJSONBModel
+from chord_metadata_service.common.base_pydantic_jsonb import AbstractPydanticJSONBModel
 
 
 __all__ = ["Project", "Dataset", "ProjectJsonSchema", "DatasetV2", "DatasetV2Translation", "DatasetV2ScopeAdapter"]
